+++ app/src/test/java/org/simple/clinic/contactpatient/ContactPatientInitTest.kt
@@ -8,7 +8,7 @@ import com.spotify.mobius.test.InitSpec.assertThatFirst
 import org.junit.Test
 import org.simple.clinic.TestData
 import org.simple.clinic.facility.FacilityConfig
-import org.simple.clinic.medicalhistory.Answer
+import org.simple.clinic.overdue.Appointment.Status.Scheduled
 import org.simple.clinic.overdue.AppointmentConfig
 import org.simple.clinic.overdue.TimeToAppointment
 import org.simple.clinic.overdue.TimeToAppointment.Days
@@ -53,17 +53,11 @@ class ContactPatientInitTest {
             teleconsultationEnabled = false
         )
     )
-    val overdueAppointment = TestData.overdueAppointment(
+    val overdueAppointment = TestData.appointment(
+        uuid = UUID.fromString("43cbf33f-102c-45e7-86b2-ea0e7601f45e"),
         facilityUuid = UUID.fromString("a607a97f-4bf6-4ce6-86a3-b266059c7734"),
         patientUuid = patientUuid,
-        patientAddress = TestData.overduePatientAddress(
-            streetAddress = null,
-            colonyOrVillage = null,
-            district = "Bhatinda",
-            state = "Punjab"),
-        patientRegisteredFacilityName = "Bhatinda",
-        diagnosedWithDiabetes = Answer.Yes,
-        diagnosedWithHypertension = Answer.No
+        status = Scheduled
     )
 
     val model = defaultModel()